ZBUG-1315: An issue with data type range values for Mysql database tables

+++ store/src/java/com/zimbra/cs/mailbox/Folder.java
@@ -1691,7 +1691,7 @@ public class Folder extends MailItem implements FolderStore {
      **/
     @Override
     public int getUIDValidity() {
-        return Math.max(getSavedSequence(), 1);
+        return (int) Math.max(getSavedSequence(), 1);
     }
 
     @Override
